move crypto import stuff into separate file and add cryptography package support.
